fixed(SchedulePagination): fixed alpha, seperate functions and helpers
