+++ components/molecules/charts/schedule/helper/prepareHiddenColumn.js
@@ -1,0 +1,7 @@
+export default (hiddenColumn) => {
+  const result = {}
+  hiddenColumn.forEach((item) => {
+    result[item] = true
+  })
+  return result
+}
